remove typings + fix runtime
