+++ packages/studio-be/src/sdk/cms/enums.ts
@@ -1,0 +1,5 @@
+export enum ButtonAction {
+  SaySomething = 'Say something',
+  OpenUrl = 'Open URL',
+  Postback = 'Postback'
+}
