+++ packages/studio-be/src/sdk/logger/enums.ts
@@ -1,0 +1,13 @@
+export enum LoggerLevel {
+  Info = 'info',
+  Warn = 'warn',
+  Error = 'error',
+  Critical = 'critical',
+  Debug = 'debug'
+}
+
+export enum LogLevel {
+  PRODUCTION = 0,
+  DEV = 1,
+  DEBUG = 2
+}
