--- packages/typings/.gitignore
@@ -1,6 +1,0 @@
-node_modules/
-yarn-error.log
-dist
-lib/
-tsconfig.tsbuildinfo
-.yarn
