--- packages/typings/global.d.ts
@@ -1,3 +1,0 @@
-declare interface Dic<T> {
-  [Key: string]: T
-}
