--- packages/typings/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "sourceMap": true,
-    "composite": true,
-    "strict": true,
-    "module": "commonjs",
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "target": "es2019",
-    "noImplicitAny": false,
-    "moduleResolution": "node",
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "outDir": "lib",
-    "types": ["reflect-metadata", "jest"]
-  },
-  "include": ["**/*.d.ts"]
-}
